fixing nock tests.
It turns out that you need to specify the query string with nock. You
can do this with the `.query()` method by passing in the literal string,
and object with the query params, or just the boolean literal `true` to
indicate that there should be a query string but you don't care what it
is.

Also the nock need to happen before the test where it's being used, so
you have to make the call insite the `it` callback or in a `beforeEach`
in the same describe block as your `it`